New lambda implementation.
The prior version requires not only a lot more code but was showing
constructors of lambdas. Technically this was incorrect, because the
standard requires direct initialization of all captures.

The new implementation swallows the pill making the caputures public and
by that get the aggregate behavior.

GCC does something similiar in its implementation. They create an
aggregate first and later switch it to a class:

https://github.com/gcc-mirror/gcc/blob/62a80185db84f20f3efb05c81598bffa95bcd63d/gcc/cp/lambda.cc#L121